requirements.txt
akshare
pandas
scipy
matplotlib
mini-racer
yfinance
numpy
fredapi
wbgapi
requests
python-dotenv